Refactored list and new components to separate packages.
