+++ src/project/Project.present.js
@@ -28,13 +28,12 @@
 import React, { Component } from 'react';
 
 import { Link, Route, Switch }  from 'react-router-dom';
-import Autosuggest from 'react-autosuggest';
 import ReactMarkdown from 'react-markdown'
 
 import { Container, Row, Col } from 'reactstrap';
-import { Alert, Badge, Table } from 'reactstrap';
+import { Alert, Table } from 'reactstrap';
 import { Button, Form, FormGroup, FormText, Label } from 'reactstrap';
-import { Input, InputGroup, InputGroupAddon, InputGroupText } from 'reactstrap';
+import { Input } from 'reactstrap';
 import { Nav, NavItem } from 'reactstrap';
 import { Card, CardBody, CardHeader } from 'reactstrap';
 
@@ -43,10 +42,9 @@ import faStarRegular from '@fortawesome/fontawesome-free-regular/faStar'
 import faStarSolid from '@fortawesome/fontawesome-free-solid/faStar'
 import faExclamationCircle from '@fortawesome/fontawesome-free-solid/faExclamationCircle'
 
-import collection from 'lodash/collection';
-import { Avatar, ExternalLink, FieldGroup, Loader, Pagination, RenkuNavLink, TimeCaption } from '../utils/UIComponents'
+import { ExternalLink, Loader, RenkuNavLink, TimeCaption } from '../utils/UIComponents'
 import { SpecialPropVal } from '../model/Model'
-import './Project.style.css';
+import { ProjectTags, ProjectTagList } from './shared'
 
 const imageBuildStatusText = {
   failed: 'No notebook image has been built. You can still open a notebook server with the default image.',
@@ -72,228 +70,6 @@ function isRequestPending(props, request) {
   const transient = props.transient || {};
   const requests = transient.requests || {}
   return requests[request] === SpecialPropVal.UPDATING;
-}
-
-class ProjectPath extends Component {
-  constructor(props) {
-    super(props);
-
-    this.onChange = this.doChange.bind(this);
-    this.onBlur = this.doBlur.bind(this);
-    this.onSuggestionsFetchRequested = this.doSuggestionsFetchRequested.bind(this);
-    this.onSuggestionsClearRequested = this.doSuggestionsClearRequested.bind(this);
-
-    this.state = {
-      suggestions: [],
-      input: this.props.namespace.name
-    };
-  }
-
-  componentDidUpdate(prevProps, prevState, snapshot) {
-    if (this.props.namespace.name !== prevProps.namespace.name) {
-      this.setState({input: this.props.namespace.name});
-    }
-  }
-
-  getSuggestionValue(suggestion) {
-    return suggestion;
-  }
-
-  renderSuggestion(suggestion) {
-    return <span>{suggestion.name}</span>;
-  }
-
-  renderSectionTitle(section) {
-    return (
-      <strong>{section.kind}</strong>
-    );
-  }
-
-  getSectionSuggestions(section) {
-    return section.namespaces;
-  }
-
-  doBlur(event, {highlightedSuggestion}) {
-    // Take the highlighted suggestion or return to the selected namespace
-    if (null != highlightedSuggestion) {
-      this.props.onChange(highlightedSuggestion);
-      this.props.onAccept();
-    }
-    this.setState({input: this.props.namespace.name})
-  }
-
-  doChange(event, { newValue, method }) {
-    // If the user typed, store it as local input, otherwise set the selection
-    if (method === 'type') {
-      this.setState({input: newValue});
-    } else {
-      this.props.onChange(newValue)
-    }
-    if (method === 'enter' || method === 'click') {
-      this.props.onAccept();
-    }
-  }
-
-  async doSuggestionsFetchRequested({ value, reason }) {
-    let searchValue = value;
-    // Do a broad search on input-focused to show many options
-    if (reason === 'input-focused') searchValue = '';
-    const matches = await this.props.fetchMatchingNamespaces(searchValue);
-    const namespaces = collection.groupBy(matches, 'kind');
-    const suggestions = ['user', 'group']
-      .map(section => {
-        const sectionNs = namespaces[section];
-        return {
-          kind: section,
-          namespaces: (sectionNs == null) ? [] : sectionNs
-        };
-      })
-      .filter(section => section.namespaces.length > 0);
-
-    // Show current as first on input-focused, otherwise do not show
-    if (reason === 'input-focused') {
-      const current = {kind: 'current', namespaces: [this.props.namespace]};
-      suggestions.splice(0, 0, current);
-    }
-    this.setState({ suggestions });
-  }
-
-  doSuggestionsClearRequested() {
-    this.setState({
-      suggestions: [],
-    });
-  }
-
-  render() {
-    const { suggestions } = this.state;
-    const inputProps = {
-      placeholder: "{user}",
-      value: this.state.input || '',
-      onChange: this.onChange,
-      onBlur: this.onBlur
-    };
-    // See https://github.com/moroshko/react-autosuggest#themeProp
-    const defaultTheme = {
-      container:                'react-autosuggest__container',
-      containerOpen:            'react-autosuggest__container--open',
-      input:                    'react-autosuggest__input',
-      inputOpen:                'react-autosuggest__input--open',
-      inputFocused:             'react-autosuggest__input--focused',
-      suggestionsContainer:     'react-autosuggest__suggestions-container',
-      suggestionsContainerOpen: 'react-autosuggest__suggestions-container--open',
-      suggestionsList:          'react-autosuggest__suggestions-list',
-      suggestion:               'react-autosuggest__suggestion',
-      suggestionFirst:          'react-autosuggest__suggestion--first',
-      suggestionHighlighted:    'react-autosuggest__suggestion--highlighted',
-      sectionContainer:         'react-autosuggest__section-container',
-      sectionContainerFirst:    'react-autosuggest__section-container--first',
-      sectionTitle:             'react-autosuggest__section-title'
-    };
-    // Override the input theme to match our visual style
-    const theme = {...defaultTheme, ...{input: 'form-control'}};
-    return <FormGroup>
-      <Label>Project Home</Label>
-      <InputGroup>
-        <Autosuggest
-          multiSection={true}
-          highlightFirstSuggestion={true}
-          suggestions={suggestions}
-          inputProps={inputProps}
-          theme={theme}
-          shouldRenderSuggestions={(v) => true}
-          onSuggestionsFetchRequested={this.onSuggestionsFetchRequested}
-          onSuggestionsClearRequested={this.onSuggestionsClearRequested}
-          getSuggestionValue={this.getSuggestionValue}
-          renderSuggestion={this.renderSuggestion}
-          renderSectionTitle={this.renderSectionTitle}
-          getSectionSuggestions={this.getSectionSuggestions} />
-        <InputGroupAddon addonType="append">
-          <InputGroupText>/</InputGroupText>
-          <InputGroupText>{this.props.slug}</InputGroupText>
-        </InputGroupAddon>
-      </InputGroup>
-      <FormText color="muted">{"By default, a project is owned by the user that created it, but \
-        it can optionally be created within a group."}</FormText>
-    </FormGroup>
-  }
-}
-
-class DataVisibility extends Component {
-  render() {
-    const visibilities = this.props.visibilities;
-    const vizExplanation = (visibilities.length === 3) ?
-      "" :
-      "The project home's visibility setting limits the project visibility options.";
-    const options = visibilities.map(v =>
-      <option key={v.value} value={v.value}>{v.name}</option>
-    )
-    return <FormGroup>
-      <Label>Visibility</Label>
-      <Input type="select" placeholder="visibility" value={this.props.value} onChange={this.props.onChange}>
-        {options}
-      </Input>
-      <FormText color="muted">{vizExplanation}</FormText>
-    </FormGroup>
-  }
-}
-
-class ProjectNew extends Component {
-
-  render() {
-    const statuses = this.props.statuses;
-    const titleHelp = this.props.model.display.slug.length > 0 ? `Id: ${this.props.model.display.slug}` : null;
-    return [
-      <Row key="header"><Col md={8}>
-        <h1>New Project</h1>
-      </Col></Row>,
-      <Row key="body"><Col md={8}>
-        <form action="" method="post" encType="multipart/form-data" id="js-upload-form">
-          <FieldGroup id="title" type="text" label="Title" placeholder="A brief name to identify the project"
-            help={titleHelp} value={this.props.model.display.title}
-            feedback={statuses.title} invalid={statuses.title != null}
-            onChange={this.props.handlers.onTitleChange} />
-          <FieldGroup id="description" type="textarea" label="Description" placeholder="A description of the project"
-            help="A description of the project helps users understand it and is highly recommended."
-            feedback={statuses.description} invalid={statuses.description != null}
-            value={this.props.model.display.description} onChange={this.props.handlers.onDescriptionChange} />
-          <ProjectPath namespace={this.props.model.meta.projectNamespace} slug={this.props.model.display.slug}
-            namespaces={this.props.namespaces}
-            onChange={this.props.handlers.onProjectNamespaceChange}
-            onAccept={this.props.handlers.onProjectNamespaceAccept}
-            fetchMatchingNamespaces={this.props.handlers.fetchMatchingNamespaces} />
-          <DataVisibility
-            value={this.props.model.meta.visibility}
-            visibilities={this.props.visibilities}
-            onChange={this.props.handlers.onVisibilityChange} />
-          <br/>
-          <Button color="primary" onClick={this.props.handlers.onSubmit}>
-            Create
-          </Button>
-        </form>
-      </Col></Row>
-    ]
-  }
-}
-
-class ProjectTag extends Component {
-  render() {
-    return <span><Badge color="primary">{this.props.tag}</Badge>&nbsp;</span>;
-  }
-}
-
-function sortedTagList(taglistOrNull) {
-  const taglist = taglistOrNull || [];
-  const tlSet = new Set(taglist);
-  const tl = Array.from(tlSet);
-  tl.sort();
-  return tl;
-}
-
-class ProjectTagList extends Component {
-  render() {
-    const taglist = sortedTagList(this.props.taglist);
-    return (taglist.length > 0) ? taglist.map(t => <ProjectTag key={t} tag={t} />) : <br />;
-  }
 }
 
 class ImageBuildInfo extends Component {
@@ -765,46 +541,6 @@ class RepositoryUrls extends Component {
   }
 }
 
-class ProjectTags extends Component {
-  constructor(props) {
-    super(props);
-    this.state = ProjectTags.getDerivedStateFromProps(props, {});
-    this.onValueChange = this.handleChange.bind(this);
-    this.onSubmit = this.handleSubmit.bind(this);
-  }
-
-  static tagListString(props) {
-    const tagList = sortedTagList(props.tag_list)
-    return tagList.join(', ');
-  }
-
-  static getDerivedStateFromProps(nextProps, prevState) {
-    const update = {value: ProjectTags.tagListString(nextProps) };
-    return {...prevState, ...update};
-  }
-
-  handleChange(e) { this.setState({value: e.target.value}); }
-
-  handleSubmit(e) { e.preventDefault(); this.props.onProjectTagsChange(this.state.value); }
-
-  render() {
-    const inputField = this.props.settingsReadOnly ?
-      <Input readOnly value={this.state.value} /> :
-      <Input value={this.state.value} onChange={this.onValueChange} />;
-    let submit = (ProjectTags.tagListString(this.props) !== this.state.value) ?
-      <Button color="primary">Update</Button> :
-      <span></span>
-    return <Form onSubmit={this.onSubmit}>
-      <FormGroup>
-        <Label for="project_tags">Project Tags</Label>
-        {inputField}
-        <FormText>Comma-separated list of tags</FormText>
-      </FormGroup>
-      {submit}
-    </Form>
-  }
-}
-
 class ProjectDescription extends Component {
   constructor(props) {
     super(props);
@@ -884,52 +620,5 @@ class ProjectView extends Component {
   }
 }
 
-class ProjectListRow extends Component {
-  render() {
-    const projectsUrl = this.props.projectsUrl;
-    const title =
-      <Link to={`${projectsUrl}/${this.props.id}`}>
-        {this.props.path_with_namespace || 'no title'}
-      </Link>
-    const description = this.props.description !== '' ? this.props.description : 'No description available';
-    return (
-      <Row className="project-list-row">
-        <Col md={2} lg={1}><Avatar person={this.props.owner} /></Col>
-        <Col md={10} lg={11}>
-          <p><b>{title}</b>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<ProjectTagList taglist={this.props.tag_list} /></p>
-          <p>{description} <TimeCaption caption="Updated" time={this.props.last_activity_at} /> </p>
-        </Col>
-      </Row>
-    );
-  }
-}
-
-class ProjectList extends Component {
-  render() {
-    const loading = this.props.loading || false;
-    const projects = this.props.page.projects || [];
-    const hasUser = this.props.user && this.props.user.id != null;
-    const rows = projects.map((d, i) => <ProjectListRow key={i} projectsUrl={this.props.urlMap.projectsUrl} {...d} />);
-    const projectsCol = (projects.length < 1 && loading) ?
-      <Col md={{size: 2,  offset: 3}}><Loader /></Col> :
-      <Col md={8}>{rows}</Col>
-    return [
-      <Row key="header">
-        <Col md={3} lg={2}><h1>Projects</h1></Col>
-        <Col md={2}>
-          {
-            (hasUser) ?
-              <Link className="btn btn-primary" role="button" to={this.props.urlMap.projectNewUrl}>New Project</Link> :
-              <span></span>
-          }
-        </Col>
-      </Row>,
-      <Row key="spacer"><Col md={8}>&nbsp;</Col></Row>,
-      <Row key="projects">{projectsCol}</Row>,
-      <Pagination key="pagination" {...this.props} />
-    ]
-  }
-}
-
-export default { ProjectNew, ProjectView, ProjectList };
-export { ProjectListRow, filterPaths };
+export default { ProjectView };
+export { filterPaths };
